+++ docs/changelog.md
@@ -2,6 +2,8 @@
 
 ## Current
 
+* feat(mail): Adds new config to send custom `from` and `subject` for LDAP reset
+password emails.
 * feat(mail): Adds new config to send custom `from` and specify a template to be
 used as an email body when resetting a password, or confirming a new account.
 The URL for mailer has also been made configurable. (ASG-2630)